+++ src/game/game_globals.h
@@ -1,0 +1,3 @@
+// Game Specific
+extern unsigned char * noise_map;
+extern int camera_x;
