+++ src/game/graphics/game_graphics.h
@@ -1,0 +1,11 @@
+/**
+ * @brief Loads textures needed for the terrain
+ * 
+ */
+void load_terrain_textures();
+
+/**
+ * @brief Draws the background terrain of the game
+ * 
+ */
+void drawTerrain();
