+++ src/game/object_handler/animation_handler.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "object_handler.h"
+#include "object/object_handler.h"
 #include "image/animation.h"
 
 class AnimationHandler : public ObjectHandler
