+++ src/game/object_handler/enemy_handler.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "object_handler.h"
+#include "object/object_handler.h"
 
 class EnemyHandler : public ObjectHandler
 {
